Refactor: - Change books to movies - Seperate containers and components
